+++ CHANGELOG.md
@@ -8,7 +8,7 @@
  * add `updatereset`. Do you need to reinitialize the camera preview when calling didupdatewidget
  * Update [`fl_camera`](https://pub.dev/packages/fl_camera)
 ## [1.0.1]
- * fix the problem that there is no data in the boundingbox
+ * fix the problem that there is no data in the `boundingbox`
 ## [1.0.0]
  * upgrade Android Gradle
 ## [0.0.6]